cut(fixtures): org lifecycle e2e → cloud-required seam e2e (kill-list A5, 2/4)
